r=float(input()) a=3.14*r*r print(f"{a:.2f}")

+++ Area_of_a_circle.py
@@ -1,3 +1,3 @@
-r=int(input())
+r=float(input())
 a=3.14*r*r
 print(f"{a:.2f}")
